Make monit dependency optional (off by default)
There's an issue due to `ansible_domain` being blank. The default address is broken otherwise. This stalls the wordpress playbook which prevents Varnish from restarting.

We can set `ansible_domain` using the bootstrap playbook. That said, you need to set `bootstrap_domain` for it to be set. This is an extra that I don't want to force on users.

Monit was added due to memory issues during initial testing. There hasn't been any issues since the `swap` role was added. This is is why I'm making turning the dependency off by default.